+++ app/src/androidTest/java/nextstep/signup/SignupButtonTest.kt
@@ -4,7 +4,7 @@ import androidx.compose.ui.test.junit4.createComposeRule
 import androidx.compose.ui.test.onNodeWithText
 import androidx.compose.ui.test.performClick
 import junit.framework.TestCase.assertTrue
-import nextstep.signup.ui.SignupButton
+import nextstep.signup.ui.component.SignupButton
 import org.junit.Rule
 import org.junit.Test
